pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cupyram"
version = "0.1.0"
description = "GPU-accelerated Range-dependent Acoustic Model (RAM) using CuPy"
readme = "README.md"
license = {text = "BSD-3-Clause"}
requires-python = ">=3.8"
authors = [
    {name = "[YOUR NAME HERE]", email = "[YOUR EMAIL HERE]"},
]
keywords = ["acoustics", "underwater", "parabolic-equation", "gpu", "cuda", "ram"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
    "Environment :: GPU :: NVIDIA CUDA",
]

dependencies = [
    "numpy>=1.20.0",
    "cupy-cuda12x>=13.0.0",
    "scipy>=1.7.0",
    "numba>=0.54.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "pyram",  # Ground truth reference implementation
]

[project.urls]
Homepage = "https://github.com/[YOUR-USERNAME]/cupyram"
Repository = "https://github.com/[YOUR-USERNAME]/cupyram"
Issues = "https://github.com/[YOUR-USERNAME]/cupyram/issues"

[tool.hatch.build.targets.sdist]
include = [
    "/cupyram",
    "/tests",
]

[tool.hatch.build.targets.wheel]
packages = ["cupyram"]

